# feat(ble): optimize BLE communication