# BLE Communication Optimization

## Implemented Optimizations

### 1. BLE 5.0 2Mbps Mode Activation
- Enabled BLE 5.0's high-speed 2Mbps physical layer rate for faster data transmission
- Implemented using `ble_gap_set_prefered_default_le_phy()` function to maximize throughput

### 2. Connection Parameter Optimization
- Reduced connection interval (7.5ms-15ms) for higher throughput
- Set slave latency to 0 for more responsive connections
- Optimized supervision timeout (1s) for reliable connections

### 3. SDK Configuration Optimization
- Increased MTU size to 512 bytes (from 256 bytes)
- Expanded memory block sizes and counts
- Optimized transport settings for better performance
- Increased GATT process count for improved handling

All optimizations have been tested and verified to build successfully with PlatformIO.